+++ Tests/E2E/.nvmrc
@@ -1,1 +1,1 @@
-v24.14.0
+v24.14.1
