+++ Tests/E2E/global-teardown.ts
@@ -4,7 +4,7 @@ import { dirname  } from 'node:path';
 export default async function globalTeardown() {
   const sut = process.env.SUT || 'neos8';
   execSync(
-    `docker compose -f ../sytem_under_test/${sut}/docker-compose.yaml down -v`,
+    `docker compose -f ../system_under_test/${sut}/docker-compose.yaml down -v`,
     { stdio: 'inherit', cwd: dirname('.') }
   );
 }
