+++ Tests/E2E/steps/hooks.ts
@@ -1,4 +1,3 @@
-import { expect } from '@playwright/test';
 import { createBdd } from 'playwright-bdd';
 import { state } from "../helpers/state.ts";
 import { logout, removeAllUsers } from "../helpers/system.ts";
@@ -8,9 +7,8 @@ const { AfterScenario } = createBdd();
 // reset db after each scenario
 AfterScenario(async ({ page }) => {
   await logout(page)
+  removeAllUsers();
 
   // clear state
   state.deviceNameSecretMap.clear();
-
-  removeAllUsers();
 });
