+++ Tests/system_under_test/neos8/docker-compose.yaml
@@ -1,0 +1,19 @@
+include:
+  - path:
+    - ../sut-base-docker-compose.yaml
+
+services:
+  neos:
+    build:
+      args:
+        PHP_VERSION: '8.2'
+        NEOS_VERSION: '8'
+        ENTRY_POINT_FILE: 'Tests/system_under_test/neos8/entrypoint.sh'
+
+volumes:
+  db_neos_data:
+    name: db_neos8_data
+
+networks:
+  neos_SUT:
+    name: neos8_SUT
