+++ Tests/system_under_test/neos9/docker-compose.yaml
@@ -1,0 +1,22 @@
+include:
+  - path:
+    - ../sut-base-docker-compose.yaml
+
+services:
+  neos:
+    build:
+      args:
+        PHP_VERSION: '8.5'
+        NEOS_VERSION: '9'
+        ENTRY_POINT_FILE: 'Tests/system_under_test/neos9/entrypoint.sh'
+
+  db:
+    image: mariadb:11.4
+
+volumes:
+  db_neos_data:
+    name: db_neos9_data
+
+networks:
+  neos_SUT:
+    name: neos9_SUT
