+++ Tests/system_under_test/sut_file_system_overrides/etc/bash.vips-arm64-hotfix.sh
