--- Tests/sytem_under_test/neos8/docker-compose.yaml
@@ -1,57 +1,0 @@
-services:
-  neos:
-    user: www-data:www-data
-    build:
-      context: ../../../
-      dockerfile: Tests/sytem_under_test/neos8/Dockerfile
-    environment:
-      FLOW_CONTEXT: '${FLOW_CONTEXT:-Production/E2E-SUT}'
-      # DB connection
-      DB_NEOS_HOST: 'db'
-      DB_NEOS_PORT: 3306
-      DB_NEOS_USER: 'neos'
-      DB_NEOS_PASSWORD: 'neos'
-      DB_NEOS_DATABASE: 'neos'
-      # Redis connection
-      REDIS_HOST: 'redis'
-      REDIS_PORT: 6379
-      # this is safe because the neos container port is only exposed to the local interface
-      # This means that the neos container is ALWAYS accessed through the front facing Ingress
-      FLOW_HTTP_TRUSTED_PROXIES: '*'
-    networks:
-      - neos8_SUT
-    ports:
-      - 8081:8081
-    depends_on:
-      - db
-      - redis
-
-  db:
-    image: mariadb:10.11
-    restart: always
-    ports:
-      - "13306:3306"
-    networks:
-      - neos8_SUT
-    environment:
-      MARIADB_RANDOM_ROOT_PASSWORD: 'true'
-      MARIADB_DATABASE: 'neos'
-      MARIADB_USER: 'neos'
-      MARIADB_PASSWORD: 'neos'
-      MARIADB_AUTO_UPGRADE: 1
-    volumes:
-      - db_neos8_data:/var/lib/mysql
-    command: ['mysqld', '--character-set-server=utf8mb4', '--collation-server=utf8mb4_unicode_ci']
-
-  redis:
-    image: redis:7
-    restart: always
-    networks:
-      - neos8_SUT
-
-volumes:
-  db_neos8_data:
-
-networks:
-  neos8_SUT:
-    name: neos8_SUT
