--- Tests/sytem_under_test/neos9/sut-files/app/Configuration/Production/E2E-SUT/EnforceForAll/Settings.2FA.yaml
@@ -1,4 +1,0 @@
-Sandstorm:
-  NeosTwoFactorAuthentication:
-    # enforce 2FA for all users
-    enforceTwoFactorAuthentication: true
