--- Tests/sytem_under_test/neos9/sut-files/app/Configuration/Production/E2E-SUT/EnforceForProvider/Settings.2FA.yaml
@@ -1,4 +1,0 @@
-Sandstorm:
-  NeosTwoFactorAuthentication:
-    # enforce 2FA for specific authentication providers (e.g. Neos.Neos:Backend)
-    enforce2FAForAuthenticationProviders : ['Neos.Neos:Backend']
