--- Tests/sytem_under_test/neos9/sut-files/app/Configuration/Production/E2E-SUT/IssuerNameChange/Settings.2FA.yaml
@@ -1,4 +1,0 @@
-Sandstorm:
-  NeosTwoFactorAuthentication:
-    # (optional) if set this will be used as a naming convention for the TOTP. If empty the Site name will be used
-    issuerName: 'Test Issuer'
