--- Tests/sytem_under_test/neos9/sut-files/app/Configuration/Production/E2E-SUT/Policy.yaml
@@ -1,5 +1,0 @@
-roles:
-  'Neos.Neos:SecondFactorUser':
-    label: 2FA Required User
-    description: Just for testing role based 2FA enforcement
-    parentRoles: ['Neos.Neos:Editor']
