--- Tests/sytem_under_test/neos9/sut-files/app/Configuration/Production/E2E-SUT/Settings.yaml
@@ -1,26 +1,0 @@
-Neos:
-  Flow:
-    persistence:
-      backendOptions:
-        driver: 'pdo_mysql'
-        charset: 'utf8mb4'
-        host: '%env:DB_NEOS_HOST%'
-        port: '%env:DB_NEOS_PORT%'
-        password: '%env:DB_NEOS_PASSWORD%'
-        user: '%env:DB_NEOS_USER%'
-        dbname: '%env:DB_NEOS_DATABASE%'
-    cache:
-      applicationIdentifier: 'app'
-
-  Imagine:
-    driver: 'Vips'
-    enabledDrivers:
-      Vips: true
-      Gd: true
-      Imagick: true
-
-  Media:
-    image:
-      defaultOptions:
-        # The Vips driver does not support interlace
-        interlace: ~
