--- Tests/sytem_under_test/neos9/sut-files/app/Configuration/Production/Settings.yaml.example
@@ -1,19 +1,0 @@
-#                                                                        #
-# Example Settings                                                       #
-#                                                                        #
-# This file contains settings for various parts of the application.      #
-# Copy this file to Settings.yaml, and adjust as necessary.              #
-#                                                                        #
-# Please refer to the default settings file(s) or the manuals for        #
-# possible configuration options.                                        #
-#                                                                        #
-
-Neos:
-  Flow:
-    persistence:
-      # For a reference of the possible options, take a look at
-      # https://www.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html
-      backendOptions:
-        dbname: 'flow'       # adjust to your database name
-        user: ''             # adjust to your database user
-        password: ''         # adjust to your database password
