--- Tests/sytem_under_test/neos9/sut-files/etc/caretakerd.yaml
@@ -1,9 +1,0 @@
-# DOCS: https://caretakerd.echocat.org/#configuration.structure
-
-rpc:
-  enabled: true
-
-services:
-  php-fpm:
-    type: master
-    command: ['frankenphp', 'run', '--config', '/etc/frankenphp/Caddyfile']
