--- Tests/sytem_under_test/neos9/sut-files/etc/frankenphp/Caddyfile
@@ -1,41 +1,0 @@
-# The Caddyfile is an easy way to configure FrankenPHP and the Caddy web server.
-#
-# https://frankenphp.dev/docs/config
-# https://caddyserver.com/docs/caddyfile
-# https://github.com/php/frankenphp/blob/main/caddy/frankenphp/Caddyfile
-{
-	skip_install_trust
-
-	# debug
-
-	frankenphp {
-		# num_threads <num_threads> # Sets the number of PHP threads to start. Default: 2x the number of available CPUs.
-		#max_threads <num_threads> # Limits the number of additional PHP threads that can be started at runtime. Default: num_threads. Can be set to 'auto'.
-		#max_wait_time <duration> # Sets the maximum time a request may wait for a free PHP thread before timing out. Default: disabled.
-		#php_ini <key> <value> # Set a php.ini directive. Can be used several times to set multiple directives.
-	}
-}
-
-:8081 {
-	# log
-
-	root /app/Web
-	encode zstd br gzip
-
-	request_body {
-		max_size 256MB
-	}
-
-  # Block direct access to PHP files except index.php
-  @blockPhp {
-    path *.php
-    not path /index.php
-  }
-
-  handle @blockPhp {
-    respond 404
-  }
-
-
-	php_server
-}
