--- Tests/sytem_under_test/neos9/sut-files/usr/local/etc/php/conf.d/php-ini-overrides.ini
@@ -1,20 +1,0 @@
-; ==================
-; Various defaults
-; ==================
-
-; for PHP >= 8.1, disable deprecations to temporarily make Neos/Flow work
-error_reporting = E_ALL & ~E_DEPRECATED & ~E_STRICT
-memory_limit = 512M
-upload_max_filesize = 256M
-post_max_size = 256M
-date.timezone = "Europe/Berlin"
-
-; ==================
-; VIPS
-; ==================
-; required for VIPS
-ffi.enable = true
-
-; for VIPS on PHP >= 8.3 the following line is required,
-; see https://github.com/libvips/php-vips
-zend.max_allowed_stack_size = -1
